fix(host): compile the host crate on Linux and macOS
The lifecycle integration test opened a file with OpenOptions::share_mode
and marked its handle with SetHandleInformation, none of which exist off
Windows, and the test carried no cfg. Six imports it needs, and the
next_agent_event helper, were only reachable from Windows-gated tests and
so counted as unused everywhere else. That broke Tests (Linux x86_64),
Linux aarch64, Style (fmt + clippy) and both package jobs.

process_is_running compared pbi_status against libc::SZOMB as u32, which
is already u32 on macOS, so clippy::unnecessary_cast failed the macOS
smoke build.

Checked with cargo clippy -p paneflow-host --all-targets -- -D warnings
against x86_64-unknown-linux-gnu, aarch64-apple-darwin and the host
target. A full cross-target workspace run is not possible here: ring
needs a Linux C compiler.

+++ crates/paneflow-host/src/process.rs
@@ -72,7 +72,7 @@ fn process_is_running(pid: u32) -> bool {
             size,
         )
     };
-    written == size && info.pbi_status != libc::SZOMB as u32
+    written == size && info.pbi_status != libc::SZOMB
 }
 
 #[cfg(not(any(windows, target_os = "linux", target_os = "macos")))]
